moved default alias config to tls config
