+++ src/main/java/com/hivemq/configuration/reader/SecurityConfigurator.java
@@ -36,7 +36,6 @@ public class SecurityConfigurator {
                 .isEnabled());
         securityConfigurationService.setAllowRequestProblemInformation(securityConfigEntity.getAllowRequestProblemInformationEntity()
                 .isEnabled());
-        securityConfigurationService.setDefaultKeystoreAlias(securityConfigEntity.getDefaultKeystoreAlias());
     }
