+++ src/main/java/com/hivemq/security/ioc/SecurityModule.java
@@ -17,7 +17,6 @@ package com.hivemq.security.ioc;
 
 import com.hivemq.bootstrap.ioc.SingletonModule;
 import com.hivemq.bootstrap.ioc.lazysingleton.LazySingleton;
-import com.hivemq.configuration.service.SecurityConfigurationService;
 import com.hivemq.security.ssl.SslContextStore;
 import com.hivemq.security.ssl.SslFactory;
 
@@ -28,18 +27,14 @@ import java.util.concurrent.ScheduledExecutorService;
  */
 public class SecurityModule extends SingletonModule<Class<SecurityModule>> {
 
-    private final SecurityConfigurationService securityConfigurationService;
-
-    public SecurityModule(final SecurityConfigurationService securityConfigurationService) {
+    public SecurityModule() {
         super(SecurityModule.class);
-        this.securityConfigurationService = securityConfigurationService;
     }
 
     @Override
     protected void configure() {
         bind(SslFactory.class).in(LazySingleton.class);
         bind(SslContextStore.class).in(LazySingleton.class);
-        bind(SecurityConfigurationService.class).toInstance(securityConfigurationService);
 
         bind(ScheduledExecutorService.class).annotatedWith(Security.class)
                 .toProvider(SecurityExecutorProvider.class)
