+++ src/test/java/com/hivemq/security/ssl/SslFactoryTest.java
@@ -79,7 +79,7 @@ public class SslFactoryTest {
         final Logger logger = (Logger) LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
         logCapture = LogbackCapturingAppender.Factory.weaveInto(logger);
 
-        final SslContextStore sslContextStore = new SslContextStore(executorService, new SslContextFactory(null));
+        final SslContextStore sslContextStore = new SslContextStore(executorService, new SslContextFactory());
         sslFactory = new SslFactory(sslContextStore);
 
         when(socketChannel.alloc()).thenReturn(byteBufAllocator);
